chatbot/requirements.txt
flask
pgvector
psycopg2
langchain
openai
python-dotenv
